fix build error and install script

+++ containers/devonly/install-agent.sh
@@ -60,14 +60,16 @@ if [ -f "${BUILD_BINARY_PATH}" ]; then
 	sudo cp "${BUILD_BINARY_PATH}" "${AGENT_BINARY_PATH}"
 	echo "Used locally built binary from ${BUILD_BINARY_PATH}"
 else
-	BINARY_URL="https://github.com/MCO-PICCOLO/Pullpiri/releases/download/v0.0.0-test/nodeagent"
-	echo "Downloading binary from ${BINARY_URL}..."
-	curl -L -o nodeagent "${BINARY_URL}"
-	if [ $? -ne 0 ]; then
-		echo "Error: Failed to download binary from ${BINARY_URL}"
-		exit 1
+	if [ ! -f "${AGENT_BINARY_PATH}" ]; then
+		BINARY_URL="https://github.com/MCO-PICCOLO/Pullpiri/releases/download/v0.0.0-test/nodeagent"
+		echo "Downloading binary from ${BINARY_URL}..."
+		curl -L -o nodeagent "${BINARY_URL}"
+		if [ $? -ne 0 ]; then
+			echo "Error: Failed to download binary from ${BINARY_URL}"
+			exit 1
+		fi
+		sudo mv -f nodeagent "${AGENT_BINARY_PATH}"
 	fi
-	sudo mv -f nodeagent "${AGENT_BINARY_PATH}"
 fi
 sudo chmod +x "${AGENT_BINARY_PATH}"
 echo "Binary installed to ${AGENT_BINARY_PATH}"
